+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This changelog starts at version `0.20.0` and includes a selection of significant changes.
 
 ## Breaking Changes
+
+  - 0.70.0
+    - Migrated from CommonJS (require, module.exports) to ES Modules (import, export) only.
+
   - 0.60.0
     - Breaking change of naming in `getFailedClients()` method: previously, it used the `includeBlankClients` parameter, which is now renamed to `includeBlank`.
 
@@ -17,6 +21,9 @@ This changelog starts at version `0.20.0` and includes a selection of significan
 
 ## Notable Changes
 
+  - 0.70.0
+    - Migrated from CommonJS (require, module.exports) to ES Modules (import, export) only.
+  
   - 0.60.0
     - Breaking change of naming in `getFailedClients()` method: previously, it used the `includeBlankClients` parameter, which is now renamed to `includeBlank`.
     - Added following parameters to `getOnlineClients()` and `getOfflineClients()` method: `includeBlank`.